Refactor for provider bridge tests
